refactor: simplify Premium experiment integration
