+++ test/assets/monetization-measurement.test.ts
@@ -8,6 +8,7 @@ describe('monetization measurement', () => {
     const promotedContent = source('app/components/pages/posts/PromotedContent.vue')
     const aiReferral = source('app/components/pages/posts/post/PostChatWithAi.vue')
     const premium = source('app/pages/premium/index.vue')
+    const premiumOriginal = source('app/components/pages/premium/PremiumLandingOriginal.vue')
     const premiumOffer = source('app/components/pages/premium/PremiumOfferLanding.vue')
     const experiments = source('app/composables/useExperiments.ts')
     const matomo = source('app/plugins/040.matomo.client.ts')
@@ -18,10 +19,12 @@ describe('monetization measurement', () => {
     expect(promotedContent).not.toContain('vIntersectionObserver')
     expect(promotedContent).not.toContain("'Impression'")
     expect(aiReferral).not.toContain('trackEvent')
-    expect(premium).not.toContain("'Checkout Outbound'")
-    expect(premium).toContain("['trackEvent', 'Premium', 'Plan Click', interval.key]")
+    expect(premiumOriginal).not.toContain("'Checkout Outbound'")
+    expect(premiumOriginal).toContain("['trackEvent', 'Premium', 'Plan Click', interval.key]")
     expect(premiumOffer).toContain("['trackEvent', 'Premium', 'Plan Click', interval.key]")
-    expect(premium).toContain('id="pricing"')
+    expect(premiumOriginal).toContain('id="pricing"')
+    expect(premium).toContain('<LazyPremiumLandingOriginal')
+    expect(premium).toContain('<LazyPremiumOfferLanding')
     expect(premiumOffer).toContain('checkoutPrice: 59.64')
     expect(premiumOffer).toContain('selectedPaymentInterval.checkoutPrice')
     expect(experiments).not.toContain('sessionStorage')
@@ -31,7 +34,9 @@ describe('monetization measurement', () => {
     expect(matomo).toContain('onNuxtReady(() =>')
     expect(matomo).toContain('trackPageView(_paq, router.currentRoute.value.fullPath, premiumLandingVariation)')
     expect(matomo).toContain("'AbTesting::create'")
-    expect(matomo.indexOf("'AbTesting::create'")).toBeLessThan(matomo.indexOf("['trackPageView']"))
+    expect(matomo.indexOf('loadAbTesting(_paq, premiumLandingVariation)')).toBeLessThan(
+      matomo.indexOf("['trackPageView']")
+    )
     expect(source('app/components/pages/home/Newsletter.vue')).not.toContain("'Newsletter', 'Submit'")
   })
 })
